requirements.txt
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
distro==1.9.0
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
drf-spectacular==0.29.0
et_xmlfile==2.0.0
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.4
metatrader5==5.0.5640
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pypdfium2==5.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-pptx==1.0.2
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.10
requests==2.33.1
rpds-py==0.30.0
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
xlsxwriter==3.2.9
